requirements.txt
attrs==21.4.0
bson==0.5.10
click==8.1.3
flake8==4.0.1
flake8-bugbear==22.4.25
flake8-builtins==1.5.3
flake8-polyfill==1.0.2
Flask==2.1.2
Flask-Admin==1.6.0
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.0
Flask-Login==0.6.1
flask-swagger-ui==4.11.1
Flask-WTF==1.0.1
greenlet==1.1.2
importlib-metadata==4.11.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
packaging==21.3
passlib==1.7.4
pep8-naming==0.12.1
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==1.4.36
tomli==2.0.1
Werkzeug==2.0.3
WTForms==3.0.1
zipp==3.8.0
